requirements.txt
Django>=4.2.0
google-cloud-pubsub>=2.18.0
google-auth>=2.17.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
python-dotenv>=1.0.0 